fix: refactor collection organization
